Separate joining and grouping functionality
